fix: Correction of bugs
Correction of a bug in population module when it calls avoid_agents method.
Correction in neighbors module so as to return a no empty DataFrame.

chore: Implemetation of the npartition parameter in the methods that can be runing with Dask.

fix: Correction of bugs

Correction of a bug in population module when it calls avoid_agents method.
Correction in neighbors module so as to return a no empty DataFrame.

fix: Correction of bugs